requirements.txt
Django>=4.2
psycopg2-binary
python-dotenv
qrcode
requests
Pillow
django-jazzmin
